Update utils.py
Honestly, I'm just mickey mousing this and hoping it works at this point. Deleting extent caused the plots to overlay each other in the top row, leaving the bottom row blank. Going to try formatting extent so that it uses array indices instead of trying to get it frequencies explicitly.